requirements.txt
defusedxml==0.4.1
Django==1.8.6
django-activeurl==0.1.9
django-allauth==0.24.1
django-cachalot==1.1.0
django-debug-toolbar==1.4
django-flat-theme==1.1.1
django-mptt==0.7.4
django-post-office==2.0.3
django-taggit==0.17.3
django-vanilla-views==1.0.4
django-widget-tweaks==1.4.1
jsonfield==1.0.3
oauthlib==1.0.3
python3-openid==3.0.7
requests==2.8.1
requests-oauthlib==0.5.0
sqlparse==0.1.18
